Moved new_elasticsearch into util
